fix(derived-input): coerce the model input frame to float64 so model_native can actually score
The null_input_policy fix was incomplete and my own end-to-end check caught it. A causal
snapshot carrying a Python None produces an OBJECT-dtype column, and LightGBM refuses those
outright ("pandas dtypes must be int, float or bool"). So `model_native` raised instead of
scoring: fail-closed rather than silent, but it did not do the job.

Coercing the single-row frame to float64 turns None into NaN, which is what native
missing-value handling actually consumes. Nothing is fabricated -- a non-numeric input raises
on the coercion rather than being silently converted.

+++ research_workflow/external_model_scoring.py
@@ -292,9 +292,14 @@ class FrozenExternalModelScorer:
             raise ExternalModelScoringError(
                 f"EXTERNAL_SCORE_INPUT_NOT_AVAILABLE_AT_CHECKPOINT: available_at_ns={available_at_ns} > checkpoint_ts={checkpoint_ts}"
             )
+        # float64, not object: a snapshot carrying a Python ``None`` yields an object-dtype column,
+        # and LightGBM refuses those outright ("pandas dtypes must be int, float or bool") -- so a
+        # model_native score over a missing input would raise instead of scoring. Coercing turns
+        # None into NaN, which is what native missing-value handling actually consumes. Nothing is
+        # fabricated: a non-numeric input raises here rather than being silently coerced.
         frame = pd.DataFrame(
             [[causal_snapshot[name] for name in features]], columns=features
-        )
+        ).astype("float64")
         null_inputs = int(frame.isna().to_numpy().sum())
         if null_inputs and self.spec.null_input_policy != "model_native":
             raise ExternalModelScoringError("external score input contains null values")
